feat: halley 조회 API, galley 조회 API 구현

+++ backend/src/main/java/org/ssafy/d210/halleyGalley/entity/Mission.java
@@ -16,19 +16,19 @@ public class Mission extends OnlyCreatedTime {
     private Long id;
 
     @Column(name = "exercise_minute", nullable = false)
-    private Integer exerciseMinute;
+    private Long exerciseMinute;
 
     @Column(name = "period", nullable = false)
     private Integer period;
 
     @Builder
-    private Mission(Integer exerciseMinute, Integer period){
+    private Mission(Long exerciseMinute, Integer period){
 
         this.exerciseMinute = exerciseMinute;
         this.period = period;
     }
 
-    public static Mission of(Integer exerciseMinute, Integer period){
+    public static Mission of(Long exerciseMinute, Integer period){
 
         return builder()
                 .exerciseMinute(exerciseMinute)
